docker_images/without-argo-events/requirements.txt
asyncio-nats-client==0.11.4
asyncio-nats-streaming==0.4.0
ed25519==1.5
nkeys==0.1.0
protobuf==3.17.3
six==1.16.0
